+++ modules/features/account/src/main/java/au/com/shiftyjelly/pocketcasts/account/di/AuthPhoneModule.kt
@@ -2,6 +2,7 @@ package au.com.shiftyjelly.pocketcasts.account.di
 
 import au.com.shiftyjelly.pocketcasts.account.watchsync.WatchSyncAuthData
 import au.com.shiftyjelly.pocketcasts.account.watchsync.WatchSyncAuthDataSerializer
+import au.com.shiftyjelly.pocketcasts.repositories.di.ForApplicationScope
 import com.google.android.horologist.annotations.ExperimentalHorologistApi
 import com.google.android.horologist.auth.data.phone.tokenshare.TokenBundleRepository
 import com.google.android.horologist.auth.data.phone.tokenshare.impl.TokenBundleRepositoryImpl
@@ -11,7 +12,6 @@ import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 import kotlinx.coroutines.CoroutineScope
-import javax.inject.Qualifier
 
 @Module
 @InstallIn(SingletonComponent::class)
@@ -30,7 +30,3 @@ object AuthPhoneModule {
         )
     }
 }
-
-@Qualifier
-@Retention(AnnotationRetention.BINARY)
-annotation class ForApplicationScope
