+++ wear/src/main/kotlin/au/com/shiftyjelly/pocketcasts/wear/di/IsEmulator.kt
@@ -5,7 +5,3 @@ import javax.inject.Qualifier
 @Qualifier
 @Retention(AnnotationRetention.BINARY)
 annotation class IsEmulator
-
-@Qualifier
-@Retention(AnnotationRetention.BINARY)
-annotation class ForApplicationScope
