+++ wear/src/main/kotlin/au/com/shiftyjelly/pocketcasts/wear/di/WearAppModule.kt
@@ -8,20 +8,11 @@ import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.SupervisorJob
 import javax.inject.Singleton
 
 @Module
 @InstallIn(SingletonComponent::class)
 object WearAppModule {
-
-    @Singleton
-    @Provides
-    @ForApplicationScope
-    fun coroutineScope(): CoroutineScope =
-        CoroutineScope(SupervisorJob() + Dispatchers.Default)
 
     @Provides
     fun connectivityManager(
